app/api/requirements.txt
fastapi[standard] == 0.112.0
pydantic==2.8.2
azure-identity==1.19.0
pydantic-settings==2.4.0
azure-cosmos==4.9.0
fastapi-azure-auth==5.0.1
marshmallow==3.19.0
azure-ai-ml==1.19.0
sseclient-py==1.8.0
uvicorn[standard]==0.34.0
opencensus-ext-azure==1.1.14
opencensus-ext-fastapi==0.1.0
